refactor: Rename `lib/outputCommand.js`
