+++ lib/output-command.js
