+++ lib/output-command.test.js
