+++ lib/plugin.js
@@ -21,7 +21,7 @@ const { test } = require('./test');
 const { getDashboardUrl, dashboardHandler } = require('./dashboard');
 const setApiGatewayAccessLogFormat = require('./setApiGatewayAccessLogFormat');
 const paramCommand = require('./paramCommand');
-const outputCommand = require('./outputCommand');
+const outputCommand = require('./output-command');
 const isAuthenticated = require('./is-authenticated');
 const throwAuthError = require('./throwAuthError');
